use fabric chat events
